+++ packages/react/src/components/Table/TableWithLoader/index.tsx
@@ -16,6 +16,8 @@ interface TableWithLoaderProps extends ComponentProps<typeof DivContainerTableRi
   isLoading: boolean;
   minWidthTable?: string;
   variant?: TableSEQMVariants;
+  autoScroll?: boolean;
+  setAutoScroll?: (value: boolean) => void;
 }
 
 const TableWithLoader: React.FC<TableWithLoaderProps> = ({
@@ -23,13 +25,21 @@ const TableWithLoader: React.FC<TableWithLoaderProps> = ({
   columns,
   isLoading,
   variant = 'light',
+  autoScroll = false,
+  setAutoScroll = () => {},
   ...rest
   // minWidthTable,
 }) => {
   return (
     <DivContainerTableRight {...rest} $isLoading={isLoading}>
       {data?.length > 0 && !isLoading ? (
-        <TableSEQM data={data} columns={columns} variant={variant} />
+        <TableSEQM
+          data={data}
+          columns={columns}
+          variant={variant}
+          autoScroll={autoScroll}
+          setAutoScroll={setAutoScroll}
+        />
       ) : (
         <>
           {isLoading ? (