+++ packages/react/src/componentsSEQM/TableSEQM/types.ts
@@ -8,6 +8,8 @@ export interface TableSEQMProps extends ComponentProps<typeof StyledTableSEQM> {
   height?: CSSProperties['height'];
   width?: CSSProperties['width'];
   variant?: TableSEQMVariants;
+  autoScroll?: boolean;
+  setAutoScroll?: (value: boolean) => void;
 }
 
 export type TableRowSEQM = {