+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @abqm-ds/tokens
 
+## 4.0.4
+
+### Patch Changes
+
+- feat: 🎸 Melhorias na table
+
+## 4.0.3
+
+### Patch Changes
+
+- feat: 🎸 Implementação do AutoScroll na table
+
 ## 4.0.2
 
 ### Patch Changes